+++ changelog.md
@@ -15,6 +15,7 @@ Codion Change Log
 ### is.codion.framework.ui
 - TabbedDetailLayout.TabbedDetailController.selectParentTabs() bug fixed, now uses the correct indexOfComponent() instead of indexOfTabComponent().
 - TabbedApplicationLayout and TabbedDetailLayout now initialize panels using prepareComponent(), panels now contain focusable components when shown.
+- EntityPanel no longer triggers display request when activated.
 
 ## 0.18.71
 ### is.codion.common.reactive